Crate atomic_polyfill

Structs

AtomicBool

A boolean type which can be safely shared between threads.

AtomicI8

An integer type which can be safely shared between threads.

AtomicI16

An integer type which can be safely shared between threads.

AtomicI32

An integer type which can be safely shared between threads.

AtomicI64

An integer type which can be safely shared between threads.

AtomicIsize

An integer type which can be safely shared between threads.

AtomicPtr

A raw pointer type which can be safely shared between threads.

AtomicU8

An integer type which can be safely shared between threads.

AtomicU16

An integer type which can be safely shared between threads.

AtomicU32

An integer type which can be safely shared between threads.

AtomicU64

An integer type which can be safely shared between threads.

AtomicUsize

An integer type which can be safely shared between threads.

Enums

Ordering

Atomic memory orderings

Constants

ATOMIC_I8_INITDeprecatedExperimental

An atomic integer initialized to 0.

ATOMIC_I16_INITDeprecatedExperimental

An atomic integer initialized to 0.

ATOMIC_I32_INITDeprecatedExperimental

An atomic integer initialized to 0.

ATOMIC_I64_INITDeprecatedExperimental

An atomic integer initialized to 0.

ATOMIC_U8_INITDeprecatedExperimental

An atomic integer initialized to 0.

ATOMIC_U16_INITDeprecatedExperimental

An atomic integer initialized to 0.

ATOMIC_U32_INITDeprecatedExperimental

An atomic integer initialized to 0.

ATOMIC_U64_INITDeprecatedExperimental

An atomic integer initialized to 0.

ATOMIC_BOOL_INITDeprecated

An AtomicBool initialized to false.

ATOMIC_ISIZE_INITDeprecated

An atomic integer initialized to 0.

ATOMIC_USIZE_INITDeprecated

An atomic integer initialized to 0.

Functions

compiler_fence

A compiler memory fence.

fence

An atomic fence.

spin_loop_hintDeprecated

Signals the processor that it is inside a busy-wait spin-loop (“spin lock”).